+++ changelog.md
@@ -15,6 +15,9 @@ _Nothing Yet..._
   - Added `BpkImage` component
   - Added `withLazyLoading` HOC which allows `BpkImage` to be loaded only when in view
   - Added `withLoadingBehavior` HOC which allows `BpkImage` to be shown only once loaded, displaying a spinner until that point
+- react-native-bpk-component-text
+  - **New component** `BpkText` is Backpack's first React Native component! It can be used on both iOS and Android
+    for adding text to apps.
 
 ## 2017-08-25 - New unlock icon